Separate out system_domain module in domain-test-service
This is preparation for the upcoming core_domain module, this commit also come with
some minor refactoring to rename the system domain generics to be more descpritive.
