+++ domains/client/domain-executor/src/aux_schema.rs
@@ -391,7 +391,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use domain_test_service::runtime::Block;
+    use domain_test_service::system_domain_test_runtime::Block;
     use sc_client_api::backend::NewBlockState;
     use sc_client_api::{Backend, BlockImportOperation};
     use sp_core::hash::H256;
